requirements.txt
black>=19.10.b0
boto3>=1.10.12
certifi>=2019.9.11
dynamodb-json>=1.3
flask>=1.1.1
pyyaml>=5.1.2
slackclient>=2.3.1
zappa>=0.48.2
